newforms-admin: Fixed #5758 -- Added a check to ensure that an inline model formset is actually deletable before checking the deletion field. Thanks, Brian Rosner and akaihola.
git-svn-id: http://code.djangoproject.com/svn/django/branches/newforms-admin@6654 bcc190cf-cafb-0310-a4f2-bffc1f526a37

+++ django/newforms/models.py
@@ -269,7 +269,7 @@ class BaseModelFormSet(BaseFormSet):
             # update/save existing instances
             for form in self.change_forms:
                 instance = instances[form.cleaned_data[self.model._meta.pk.attname]]
-                if form.cleaned_data[DELETION_FIELD_NAME]:
+                if self.deletable and form.cleaned_data[DELETION_FIELD_NAME]:
                     instance.delete()
                 else:
                     saved_instances.append(self.save_instance(form, instance, commit=commit))
